// ui-app/src/app/models/api-entities/business-partner-profile.ts
export interface BusinessPartnerProfile {
  id: string;
  ario_marketing_requested_at: string;
  collateral_downloaded_at: string;
  first_customer_invited_at: string;
  first_customer_invoiced_at: string;
  partner_training_completed_at: string;
  sales_training_completed_at: string;
  facebook_sharing_requested_at: string;
  twitter_sharing_requested_at: string;
  linkedin_sharing_requested_at: string;
  vanity_added_to_website_at: string;
  created_at: string;
  status: string;
  merchant_id: string;
  application_id: string;
}